Pushing script.js
The script checks whether the word is a palindrome by cleaning the string to remove symbols, then putting them in lowercase (to allow the program to be case-insensitive) then reversing it and comparing it to the original to see any differences. This connects to 'index.html' to run on the webpage.